+++ sdk/communication/communication-common/src/credential/clientArguments.ts
@@ -1,18 +1,17 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { isTokenCredential, KeyCredential, TokenCredential } from "@azure/core-auth";
-import { URLBuilder } from "@azure/core-http";
+import { KeyCredential, TokenCredential, isTokenCredential } from "@azure/core-auth";
 import { parseConnectionString } from "./connectionString";
 
 const isValidEndpoint = (host: string): boolean => {
-  const url = URLBuilder.parse(host);
+  const url = new URL(host);
 
   return (
-    !!url.getScheme()?.match(/^http[s]?/) &&
-    url.getHost() !== undefined &&
-    url.getHost() !== "" &&
-    (url.getPath() === undefined || url.getPath() === "" || url.getPath() === "/")
+    !!url.protocol?.match(/^http[s]?/) &&
+    url.host !== undefined &&
+    url.host !== "" &&
+    (url.pathname === undefined || url.pathname === "" || url.pathname === "/")
   );
 };