+++ sdk/compute/arm-compute/src/operations/restorePoints.ts
@@ -324,7 +324,7 @@ const getOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.CloudError
     }
   },
-  queryParameters: [Parameters.apiVersion],
+  queryParameters: [Parameters.apiVersion, Parameters.expand4],
   urlParameters: [
     Parameters.$host,
     Parameters.resourceGroupName,