+++ sdk/eventhub/eventhubs-checkpointstore-blob/src/index.ts
@@ -3,4 +3,5 @@
 /// <reference lib="esnext.asynciterable" />
 
 export { BlobCheckpointStore } from "./blobCheckpointStore";
+export { ContainerClientLike, BlobClientLike, BlockBlobClientLike } from "./storageBlobInterfaces";
 export { logger } from "./log";