+++ sdk/recoveryservicesbackup/arm-recoveryservicesbackup/samples-dev/listWorkloadItemsInContainer.ts
@@ -14,7 +14,7 @@ parameters.
  *
  * @summary Provides a pageable list of workload item of a specific container according to the query filter and the pagination
 parameters.
- * x-ms-original-file: specification/recoveryservicesbackup/resource-manager/Microsoft.RecoveryServices/stable/2021-10-01/examples/AzureWorkload/BackupWorkloadItems_List.json
+ * x-ms-original-file: specification/recoveryservicesbackup/resource-manager/Microsoft.RecoveryServices/stable/2021-12-01/examples/AzureWorkload/BackupWorkloadItems_List.json
  */
 import { RecoveryServicesBackupClient } from "@azure/arm-recoveryservicesbackup";
 import { DefaultAzureCredential } from "@azure/identity";