+++ sdk/recoveryservicesbackup/arm-recoveryservicesbackup/samples/v8/javascript/getVaultSecurityConfig.js
@@ -12,7 +12,7 @@
  * This sample demonstrates how to Fetches resource vault config.
  *
  * @summary Fetches resource vault config.
- * x-ms-original-file: specification/recoveryservicesbackup/resource-manager/Microsoft.RecoveryServices/stable/2021-10-01/examples/Common/BackupResourceVaultConfigs_Get.json
+ * x-ms-original-file: specification/recoveryservicesbackup/resource-manager/Microsoft.RecoveryServices/stable/2021-12-01/examples/Common/BackupResourceVaultConfigs_Get.json
  */
 const { RecoveryServicesBackupClient } = require("@azure/arm-recoveryservicesbackup");
 const { DefaultAzureCredential } = require("@azure/identity");